higher half, paging semi done
this commit breaks and/or discards some of the progress on the OS. due to poor design and bad design decisions i am implementing paging and moving the kernel to the higher half now and not in the beginning. i have to push in order to be able to continue the work on my main PC instead of the laptop, on which I was coding on the go. for now you have to use the serial monitor in qemu (ctrl-alt-3) to view the logs. at least it doesn't throw page faults now lol.